@@ -17,26 +17,24 @@
   "dependencies": {
     "commandpost": "^0.1.1",
     "definition-header": "^0.1.0",
-    "es6-promise": "2.0.1",
-    "github": "^0.2.3"
+    "es6-promise": "^2.3.0",
+    "github": "^0.2.4"
   },
   "devDependencies": {
-    "espower-loader": "0.10.0",
+    "espower-loader": "0.11.0",
     "grunt": "0.4.5",
     "grunt-contrib-clean": "0.6.0",
-    "grunt-conventional-changelog": "1.2.1",
-    "grunt-dtsm": "0.2.8",
+    "grunt-dtsm": "0.2.9",
     "grunt-mocha-test": "0.12.7",
-    "grunt-ts": "4.0.1",
-    "grunt-tslint": "2.0.0",
+    "grunt-ts": "4.2.0",
+    "grunt-tslint": "2.4.0",
     "grunt-typedoc": "0.2.2",
-    "load-grunt-tasks": "3.1.0",
-    "mocha": "2.2.1",
-    "nexpect": "0.5.0",
-    "power-assert": "0.10.2",
-    "source-map-support": "0.2.10",
-    "time-grunt": "1.1.0",
-    "typedoc": "0.3.2",
-    "typescript": "1.4.1"
+    "load-grunt-tasks": "3.2.0",
+    "mocha": "2.2.5",
+    "power-assert": "0.11.0",
+    "source-map-support": "0.3.2",
+    "time-grunt": "1.2.1",
+    "typedoc": "0.3.8",
+    "typescript": "1.5.3"
   }
 }
